+++ native/src/main/java/io/ballerina/stdlib/http/transport/contractimpl/DefaultHttpClientConnector.java
@@ -60,6 +60,7 @@ import java.io.IOException;
 import java.net.InetSocketAddress;
 import java.util.Calendar;
 import java.util.NoSuchElementException;
+import java.util.Objects;
 
 import static io.ballerina.stdlib.http.transport.contract.Constants.OUTBOUND_ACCESS_LOG_MESSAGE;
 import static io.ballerina.stdlib.http.transport.contract.Constants.REMOTE_SERVER_CLOSED_BEFORE_INITIATING_OUTBOUND_REQUEST;
@@ -391,5 +392,11 @@ public class DefaultHttpClientConnector implements HttpClientConnector {
         this.socketIdleTimeout = senderConfiguration.getSocketIdleTimeout(Constants.ENDPOINT_TIMEOUT);
         this.sslConfig = senderConfiguration.getClientSSLConfig();
         this.forwardedExtensionConfig = senderConfiguration.getForwardedExtensionConfig();
+    }
+
+    public void initializeSSLContext() throws Exception {
+        if (Objects.nonNull(sslConfig)) {
+            sslConfig.initializeSSLContext(http2);
+        }
     }
 }