+++ native/src/test/java/io/ballerina/stdlib/http/transport/http2/ssl/Http2MutualSslTest.java
@@ -51,7 +51,7 @@ public class Http2MutualSslTest {
     private HttpWsConnectorFactory connectorFactory;
 
     @BeforeClass
-    public void setup() throws InterruptedException {
+    public void setup() throws Exception {
 
         HttpWsConnectorFactory factory = new DefaultHttpWsConnectorFactory();
         serverConnector = factory
@@ -62,7 +62,7 @@ public class Http2MutualSslTest {
 
         connectorFactory = new DefaultHttpWsConnectorFactory();
         http2ClientConnector = connectorFactory
-                .createHttpClientConnector(new HashMap<>(), getSenderConfigs());
+                .createHttpsClientConnector(new HashMap<>(), getSenderConfigs());
     }
 
     private SenderConfiguration getSenderConfigs() {