+++ native/src/test/java/io/ballerina/stdlib/http/transport/https/MutualSSLTestCase.java
@@ -50,7 +50,7 @@ public class MutualSSLTestCase {
     private ServerConnector connector;
 
     @BeforeClass
-    public void setup() throws InterruptedException {
+    public void setup() throws Exception {
 
         factory = new DefaultHttpWsConnectorFactory();
 
@@ -62,7 +62,7 @@ public class MutualSSLTestCase {
         future.setHttpConnectorListener(new EchoMessageListener());
         future.sync();
 
-        httpClientConnector = factory.createHttpClientConnector(new HashMap<>(), getSenderConfigs());
+        httpClientConnector = factory.createHttpsClientConnector(new HashMap<>(), getSenderConfigs());
     }
 
     private ListenerConfiguration getListenerConfiguration() {