+++ native/src/test/java/io/ballerina/stdlib/http/transport/https/SSLProtocolsWithCertsTest.java
@@ -83,7 +83,7 @@ public class SSLProtocolsWithCertsTest {
      */
     @Test(dataProvider = "protocols")
     public void setup(String clientProtocol, String serverProtocol, boolean hasException, int serverPort)
-            throws InterruptedException {
+            throws Exception {
 
         Parameter clientprotocols = new Parameter("sslEnabledProtocols", clientProtocol);
         clientParams.add(clientprotocols);
@@ -101,7 +101,7 @@ public class SSLProtocolsWithCertsTest {
         future.setHttpConnectorListener(new EchoMessageListener());
         future.sync();
 
-        httpClientConnector = httpWsConnectorFactory.createHttpClientConnector(new HashMap<>(), getSenderConfigs());
+        httpClientConnector = httpWsConnectorFactory.createHttpsClientConnector(new HashMap<>(), getSenderConfigs());
 
         testSSLProtocols(hasException, serverPort);
         serverConnector.stop();