+++ native/src/test/java/io/ballerina/stdlib/http/transport/https/ServerCloseConnectionDuringSslTest.java
@@ -57,7 +57,7 @@ public class ServerCloseConnectionDuringSslTest {
     private DefaultHttpConnectorListener listener;
 
     @BeforeClass
-    public void setup() {
+    public void setup() throws Exception {
         givenServerThatClosesConnection();
         givenANormalHttpsClient();
     }
@@ -115,9 +115,9 @@ public class ServerCloseConnectionDuringSslTest {
         return senderConfiguration;
     }
 
-    private void givenANormalHttpsClient() {
+    private void givenANormalHttpsClient() throws Exception {
         factory = new DefaultHttpWsConnectorFactory();
-        httpClientConnector = factory.createHttpClientConnector(new HashMap<>(), getSenderConfigs());
+        httpClientConnector = factory.createHttpsClientConnector(new HashMap<>(), getSenderConfigs());
     }
 
     private void givenServerThatClosesConnection() {